# swagger/swagger.json
{
	"components": {
		"schemas": {
...
			"malak.Integration": {
				"properties": {
					"created_at": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"integration_name": {
						"type": "string"
					},
					"integration_type": {
						"$ref": "#/components/schemas/malak.IntegrationType"
					},
					"is_enabled": {
						"type": "boolean"
					},
					"logo_url": {
						"type": "string"
					},
					"metadata": {
						"$ref": "#/components/schemas/malak.IntegrationMetadata"
					},
					"reference": {
						"type": "string"
					},
					"updated_at": {
						"type": "string"
					}
				},
				"type": "object"
			},
...
			"malak.IntegrationMetadata": {
				"properties": {
					"endpoint": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"malak.IntegrationType": {
				"enum": [
					"oauth2",
					"api_key"
				],
				"type": "string",
				"x-enum-varnames": [
					"IntegrationTypeOauth2",
					"IntegrationTypeApiKey"
				]
			},
...
			"malak.WorkspaceIntegration": {
				"properties": {
					"created_at": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"integration": {
						"$ref": "#/components/schemas/malak.Integration"
					},
					"integration_id": {
						"type": "string"
					},
					"is_active": {
						"description": "IsActive determines if the connection to the integration has been tested and works",
						"type": "boolean"
					},
					"is_enabled": {
						"description": "IsEnabled - this integration is enabled and data can be fetched",
						"type": "boolean"
					},
					"metadata": {
						"$ref": "#/components/schemas/malak.WorkspaceIntegrationMetadata"
					},
					"reference": {
						"type": "string"
					},
					"updated_at": {
						"type": "string"
					},
					"workspace_id": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"malak.WorkspaceIntegrationMetadata": {
				"properties": {
					"access_token": {
						"type": "string"
					},
					"last_fetched_at": {
						"type": "string"
					}
				},
				"type": "object"
			},
...
			"server.APIStatus": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object"
			},
...
			"server.listIntegrationResponse": {
				"properties": {
					"integrations": {
						"items": {
							"$ref": "#/components/schemas/malak.WorkspaceIntegration"
						},
						"type": "array"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"integrations",
					"message"
				],
				"type": "object"
			},
...
			"server.testAPIIntegrationRequest": {
				"properties": {
					"api_key": {
						"type": "string"
					}
				},
				"required": [
					"api_key"
				],
				"type": "object"
			},
...
		},
		"securitySchemes": {
			"ApiKeyAuth": {
				"in": "header",
				"name": "Authorization",
				"type": "apiKey"
			}
		}
	},
	"info": {
		"contact": {
			"email": "lanre@ayinke.ventures",
			"name": "Ayinke Ventures"
		},
		"title": "Malak's API documentation",
		"version": "0.1.0"
	},
	"openapi": "3.0.3",
	"paths": {
...
		"/workspaces/integrations": {
			"get": {
				"description": "fetch workspace preferences",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.listIntegrationResponse"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Not Found"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"integrations"
				]
			}
		},
		"/workspaces/integrations/{reference}": {
			"delete": {
				"description": "disable integration",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Not Found"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"integrations"
				]
			},
			"post": {
				"description": "enable integration",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/server.testAPIIntegrationRequest"
							}
						}
					},
					"description": "request body",
					"required": true,
					"x-originalParamName": "message"
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Not Found"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"integrations"
				]
			},
			"put": {
				"description": "update integration api key",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/server.testAPIIntegrationRequest"
							}
						}
					},
					"description": "request body",
					"required": true,
					"x-originalParamName": "message"
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Not Found"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"integrations"
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:5300/v1"
		}
	]
}